add contact page

+++ src/components/ContactItem/index.jsx
@@ -1,12 +1,13 @@
 import {useDispatch} from "react-redux";
-import {Box, Link} from "@mui/material";
+import {Link} from "react-router-dom";
+import {CancelOutlined} from "@mui/icons-material";
+import {Box, Link as MuiLink} from "@mui/material";
 import {
   ContactCard,
   ContactCardContent,
   ContactIconButton,
   ContactTypography
 } from "./styledComponents.js";
-import {CancelOutlined} from "@mui/icons-material";
 import {deleteContact, fetchContacts} from "../../store/contacts/asyncActions.js";
 import ContactAvatar from "../ContactAvatar";
 import Tags from "../Tags";
@@ -15,11 +16,13 @@ const ContactItem = ({id, avatar, name, email, tags}) => {
   const dispatch = useDispatch();
 
   const deleteContactHandler =  async () => {
-    try {
-      await dispatch(deleteContact(id));
-      dispatch(fetchContacts());
-    } catch (e) {
-      return e;
+    if (window.confirm('Do you want to remove the contact?')) {
+      try {
+        await dispatch(deleteContact(id));
+        dispatch(fetchContacts());
+      } catch (e) {
+        return e;
+      }
     }
   }
 
@@ -32,15 +35,19 @@ const ContactItem = ({id, avatar, name, email, tags}) => {
       >
         <CancelOutlined />
       </ContactIconButton>
-      <ContactAvatar avatar={avatar} name={name} />
+      <Link to={`contact/${id}`}>
+        <ContactAvatar avatar={avatar} name={name} />
+      </Link>
       <Box>
         <ContactCardContent>
-          <ContactTypography component="h3" variant="h3">
-            {name}
-          </ContactTypography>
-          <Link href={`mailto:${email}`} color="inherit">
-            {email}
+          <Link to={`contact/${id}`}>
+            <ContactTypography component="h3" variant="h3">
+              {name}
+            </ContactTypography>
           </Link>
+          <MuiLink href={`mailto:${email}`} color="inherit">
+            {email}
+          </MuiLink>
           <Tags tags={tags} />
         </ContactCardContent>
       </Box>
